Long Overdue Update
About a year later, just making this new commit. Program ended up
forfeiting many games due to spending all of its allotted time. This was
due to a lack of an interrupt in the alpha-beta search so it was
completed its tree search for whichever depth it was at, which took a
lot of time past about a depth of 4.